Refactoring - removed ReservedStreamIndex concept from ClusteredStreams - replaced with string AttachmentID - reworked clustered attachment pattern - cleaned up index's for ObjectStreams - cleaned up merkle tree indexes - other related refactorings

+++ src/Hydrogen/Collections/Lists/IExtendedList.cs
@@ -20,9 +20,9 @@ public interface IExtendedList<T> : IExtendedCollection<T>, IReadOnlyExtendedLis
 
 	new int IndexOf(T item);
 
-	new void Insert(int index, T item);
+	new void Insert(long index, T item);
 
-	new void RemoveAt(int index);
+	new void RemoveAt(long index);
 }
